feat(src): Extract downloading to download_dataset.py
* Create dvc pipeline to trigger with 'dvc repro'
* Allow separate download_dataset and main.py with train
