+++ .github/workflows/test-and-coverage.yml
@@ -10,7 +10,6 @@ permissions:
     pages: write
     id-token: write
 
-
 jobs:
 
     mypy-tests:
@@ -92,7 +91,6 @@ jobs:
             -   uses: actions/setup-python@v5
                 with:
                     python-version: "3.10"
-
             -   name: Install dependencies
                 run: |
                     python -m pip install --upgrade pip
