+++ tox.ini
@@ -43,3 +43,6 @@ deps =
     mypy
     -r {toxinidir}/requirements.txt
 commands = mypy
+
+[pytest]
+tmp_path_retention_policy = "none"
